Divided the whole code in 4 different modules.
